+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
 
+### 0.2.0
+- Added command
+	- If ChikachiLib is detected, /chikachi discord
+	- Otherwise, /discord
+- Broadcast messages (/say) is now also relayed
+- Made Discord mentions case-insensitive
+
 #### 0.1.3
 - Fix NullPointerException on startup
-- Added experimental fake players (CAN FILL YOUR SERVER UP WITH FAKE PLAYERS)
+- Added experimental fake players (CAN FILL YOUR SERVER UP WITH FAKE PLAYERS)  
 ![Example](https://ss.chikachi.net/2016-05-05_03-10-13_7109d217-ef20-4979-a5a4-0baa8c8a46c4.png)
 - Call proper shutdown of Discord connection